scripts/requirements.txt
ccxt>=4.0.0
pandas>=1.5.0
python-dateutil>=2.8.0

